crud completo

+++ blogPessoal/src/app/delete-tema/delete-tema.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Tema } from '../model/Tema';
 import { TemaService } from '../service/tema.service';
 import { Router, ActivatedRoute } from '@angular/router';
+import { AlertasService } from '../service/alertas.service';
 
 @Component({
   selector: 'app-delete-tema',
@@ -15,7 +16,8 @@ export class DeleteTemaComponent implements OnInit {
   constructor(
     private temaService: TemaService,
     private router: Router,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+    private alerta: AlertasService
   ) { }
 
   ngOnInit() {
@@ -32,10 +34,15 @@ export class DeleteTemaComponent implements OnInit {
   }
 
   btnSim(){
-    this.temaService.deleteTema(this.tema.id).subscribe((resp: Tema) =>{
+    if (this.tema.postagem.length != 0) {
+      this.alerta.showAlertDanger('Esse tema não pode ser modificado, pois já pertence a uma postagem.')
       this.router.navigate(['/cadastro-tema'])
-      alert('Tema apagado com sucesso!')
-    })
+    } else {
+      this.temaService.deleteTema(this.tema.id).subscribe((resp: Tema) =>{
+        this.router.navigate(['/cadastro-tema'])
+        this.alerta.showAlertSuccess('Tema apagado com sucesso!')
+      })
+    }
   }
 
   btnNao(){
